requirements.txt
anyio==3.6.1
autopep8==1.7.0
click==8.1.3
fastapi==0.85.0
h11==0.13.0
idna==3.4
mysql-connector-python==8.0.30
protobuf==3.20.1
pycodestyle==2.9.1
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
typing-extensions==4.3.0
uvicorn==0.18.3
